+++ src/main/java/com/majon/sns_copy/feed/FeedController.java
@@ -2,6 +2,7 @@ package com.majon.sns_copy.feed;
 
 import java.util.List;
 
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,8 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.majon.sns_copy.feed.bo.FeedBO;
-import com.majon.sns_copy.model.Comment;
-import com.majon.sns_copy.model.Feed;
+import com.majon.sns_copy.feed.model.Feed;
 
 @Controller
 @RequestMapping("/feed")
@@ -18,24 +18,25 @@ public class FeedController {
 
 	@Autowired
 	private FeedBO feedBO;
+
 	
 	@GetMapping("/main_view")
 	public String mainPage(Model model) {
 		
-		// 피드 리스트
-		List<Feed> printFeedList = feedBO.getPost();
 		
-		// comment 리스트
-		List<Comment> printCommentList = feedBO.getComment();
+		List<Feed> printFeedList = feedBO.getPost();
 		
 		if(printFeedList!=null) {
 			model.addAttribute("printFeedList", printFeedList);
 		}
 		
+		/*List<Comment> printCommentList = commentBO.getComment();
+		
 		if(printCommentList!=null) {
 			model.addAttribute("printCommentList", printCommentList);
-		}
-			
+		}	
+		*/
+		
 		return "/feed/Main";
 	}
 }
