+++ src/main/java/com/majon/sns_copy/feed/dao/FeedDAO.java
@@ -2,11 +2,11 @@ package com.majon.sns_copy.feed.dao;
 
 import java.util.List;
 
+
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-import com.majon.sns_copy.model.Comment;
-import com.majon.sns_copy.model.Feed;
+import com.majon.sns_copy.feed.model.Feed;
 
 @Repository
 public interface FeedDAO {
@@ -18,13 +18,5 @@ public interface FeedDAO {
 			,@Param("imagePath") String imagePath);
 	
 	public List<Feed> selectPost();
-	
-	public int insertComment(
-			@Param("userId")int userId
-			,@Param("feedId")int feedId
-			,@Param("userName")String userName
-			,@Param("content")String content);
-	
-	public List<Comment> selectComment();
-	
+
 }
