+++ src/main/java/com/majon/sns_copy/user/dao/UserDAO.java
@@ -3,7 +3,7 @@ package com.majon.sns_copy.user.dao;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
-import com.majon.sns_copy.model.User;
+import com.majon.sns_copy.user.model.User;
 
 @Repository
 public interface UserDAO {
